Migrate Dictionary to IOperation
