--- src/FluentAssertions.Analyzers/Tips/Dictionaries/DictionaryAnalyzer.cs
@@ -1,13 +1,0 @@
-﻿using FluentAssertions.Analyzers.Utilities;
-using Microsoft.CodeAnalysis;
-
-namespace FluentAssertions.Analyzers;
-
-public abstract class DictionaryAnalyzer : FluentAssertionsAnalyzer
-{
-    protected override bool ShouldAnalyzeVariableNamedType(INamedTypeSymbol type, SemanticModel semanticModel)
-    {
-        var iDictionaryType = semanticModel.GetGenericIDictionaryType();
-        return type.IsTypeOrConstructedFromTypeOrImplementsType(iDictionaryType);
-    }
-}
